Moved Plot extension into the extensions file and renamed Schema -> StructuredDataSchema.
