+++ Sources/App/Views/Plot+Extensions.swift
@@ -1,4 +1,10 @@
+import Foundation
 import Plot
+
+protocol StructuredDataSchema: Encodable {
+    var context: String { get }
+    var type: String { get }
+}
 
 extension Node where Context: HTML.BodyContext {
     static func turboFrame(id: String, source: String? = nil, _ nodes: Node<HTML.BodyContext>...) -> Self {
@@ -7,6 +13,25 @@ extension Node where Context: HTML.BodyContext {
             .attribute(named: "src", value: source)
         ]
         return .element(named: "turbo-frame", nodes: attributes + nodes)
+    }
+
+    static func structuredData<T>(_ model: T) -> Node<HTML.BodyContext> where T: StructuredDataSchema {
+        let encoder = JSONEncoder()
+        encoder.dateEncodingStrategy = .iso8601
+        encoder.outputFormatting = [ .sortedKeys, .withoutEscapingSlashes ]
+
+        do {
+            let data = try encoder.encode(model)
+            let rawScript = String(decoding: data, as: UTF8.self)
+
+            return .script(
+                .attribute(named: "type", value: "application/ld+json"),
+                .raw(rawScript)
+            )
+        } catch {
+            AppEnvironment.logger?.error("Failed to encode structured data model: \(error)")
+            return .empty
+        }
     }
 
     static func spiReadme(_ nodes: Node<HTML.BodyContext>...) -> Self {
